fix(migrations/logging): to none without monitoring
Fix migrations of logging from * to none when monitoring is disabled.

logging migrations to none (i.e. deleting logging resources) failed if _monitoring_ was none because we were trying to delete resources using the monitoring.coreos.com/v1 API that was not available. This fix filters out these type of resources from the deletion.