feat(on-prem): support kube-proxy-less clusters
Add support for CREATING on-premises clusters without the kubeproxy.

Choosing to disable the kubeproxy will take effect on some parts of the distro, i.e.:
- enable kube-proxy-replacement options for Cilium.
- enable eBPF Mode for Calico.
- remove the kubeproxy monitoring components.

NOTE:

Calico has been left out of this commit because at the time of this writing there are some issues of cyclic dependencies whith eBPF mode that need to be solved.

In consequence, we are setting a condition on the on-premises schema that checks that Cilium is chosen when kubeproxy is disabled.