feat(on-prem): support kube-proxy-less clusters
Add support for CREATING on-premises clusters without the kube-proxy.

Choosing to disable the kube-proxy will take effect on some parts of the distro, i.e.:
- enable kube-proxy-replacement options for Cilium.
- enable eBPF Mode for Calico.
- remove the kube-proxy monitoring components.

NOTE: Calico has been left out of this commit because at the time of this writing there are some issues of cyclic dependencies whith eBPF mode that need to be solved. See https://github.com/projectcalico/calico/issues/10683

In consequence, we are setting a condition on the on-premises schema that checks that Cilium is chosen when kube-proxy is disabled.